pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "pydantic[standard]>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
